migrate to interop-testing module
