+++ core/build.gradle
@@ -33,8 +33,7 @@ dependencies {
     testImplementation project(':grpc-context').sourceSets.test.output,
             project(':grpc-api').sourceSets.test.output,
             project(':grpc-testing'),
-            project(':grpc-grpclb'),
-            project(':grpc-netty')
+            project(':grpc-grpclb')
     testImplementation (libraries.guava_testlib) {
         exclude group: 'junit', module: 'junit'
     }
