+++ interop-testing/build.gradle
@@ -43,6 +43,7 @@ dependencies {
             libraries.netty_tcnative,
             project(':grpc-grpclb')
     testImplementation project(':grpc-context').sourceSets.test.output,
+            project(':grpc-api').sourceSets.test.output,
             libraries.mockito
     alpnagent libraries.jetty_alpn_agent
 }
